Allow building the code with Clang 10 on Linux
Summary:
Allow building the code with Clang 10 on Linux.

Notes:
- We are using the "uninstrumented" directory instead of "clang_uninstrumented" for Clang-only third-party dependency archives. This is because when everything is built with Clang, we should not use the "clang_" prefix for just one directory in $YB_THIRDPARTY_DIR/installed.
- Adding YB_GCC_SUFFIX and YB_CLANG_SUFFIX variables to allow appending a suffix like "-<version>" to the compiler path.

Test Plan: Jenkins

Reviewers: jharveysmith, steve.varnau

Reviewed By: steve.varnau

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D9677

+++ src/postgres/src/backend/utils/adt/timestamp.c
@@ -3215,7 +3215,8 @@ interval_mul(PG_FUNCTION_ARGS)
 	/* cascade units down */
 	result->day += (int32) month_remainder_days;
 	result_double = rint(span->time * factor + sec_remainder * USECS_PER_SEC);
-	if (result_double > PG_INT64_MAX || result_double < PG_INT64_MIN)
+	/* Do an explicit conversion to avoid the -Wimplicit-int-float-conversion warning */
+	if (result_double > (double) PG_INT64_MAX || result_double < (double) PG_INT64_MIN)
 		ereport(ERROR,
 				(errcode(ERRCODE_DATETIME_VALUE_OUT_OF_RANGE),
 				 errmsg("interval out of range")));
